Add NTFS support

+++ test/JavaTests/testdata/multiple-files/compress.sh
@@ -16,6 +16,7 @@ CREATE_SIMPLE_XAR=n
 CREATE_SIMPLE_UDF=n
 CREATE_SIMPLE_WIM=n
 CREATE_SIMPLE_FAT=n
+CREATE_SIMPLE_NTFS=n
 
 
 TMP=/tmp
@@ -208,32 +209,54 @@ if test $CREATE_SIMPLE_FAT = y -o $CREATE_ALL = y ;then
     mkdir -p __tmp_volume
     rm -f fat/*.fat
     rm -f fat/*.zip
-    #for i in 0:maximum 1:fast 2:none
 
     for sectorsPerCluster in 1 2 4
     do
       for i in 1:1300 2:70 3:60
       do
-          #for j in *.dat
-          #do
-              FILE="fat/archive${i:0:1}.zip.$sectorsPerCluster.fat"
-              echo ""
-              echo "----------------> Creating $FILE"
-              echo ""
-              dd if=/dev/zero "of=$FILE" count=${i:2} bs=1024 \
-                && mkfs.vfat -s $sectorsPerCluster "$FILE" \
-                && sudo mount "$FILE" __tmp_volume \
-                && sudo unzip -q archive${i:0:1}.zip -d __tmp_volume/ \
-                && ls -l __tmp_volume \
-                && echo "COPIED SUCCESSFULLY"
-              sync
-              sudo umount "$PWD/__tmp_volume"
-              zip -9 --junk-paths $FILE.zip $FILE
-              rm $FILE
-              #cp $j /$IMAGEX_SUBST_DRIVE/
-              #cmd /c "imagex /capture $IMAGEX_SUBST_DRIVE: wim/$j.${i:0:1}.wim Simple_test__$j /compress ${i:2}"
-          #done
+        FILE="fat/archive${i:0:1}.zip.$sectorsPerCluster.fat"
+        echo ""
+        echo "----------------> Creating $FILE"
+        echo ""
+        dd if=/dev/zero "of=$FILE" count=${i:2} bs=1024 \
+          && mkfs.vfat -s $sectorsPerCluster "$FILE" \
+          && sudo mount "$FILE" __tmp_volume \
+          && sudo unzip -q archive${i:0:1}.zip -d __tmp_volume/ \
+          && ls -l __tmp_volume \
+          && echo "COPIED SUCCESSFULLY"
+        sync
+        sudo umount "$PWD/__tmp_volume"
+        zip -9 --junk-paths $FILE.zip $FILE
+        rm $FILE
       done
+    done
+    rm -rf __tmp_volume
+fi
+
+
+if test $CREATE_SIMPLE_NTFS = y -o $CREATE_ALL = y ;then
+    mkdir -p ntfs
+    rm -rf __tmp_volume
+    mkdir -p __tmp_volume
+    rm -f ntfs/*.ntfs
+    rm -f ntfs/*.zip
+
+    for i in 1:1190 2:1025 3:1025 4:1025
+    do
+      FILE="ntfs/archive${i:0:1}.zip.1.ntfs"
+      echo ""
+      echo "----------------> Creating $FILE"
+      echo ""
+      dd if=/dev/zero "of=$FILE" count=${i:2} bs=1024 \
+        && mkfs.ntfs -F "$FILE" \
+        && sudo mount "$FILE" __tmp_volume \
+        && sudo unzip -q archive${i:0:1}.zip -d __tmp_volume/ \
+        && ls -l __tmp_volume \
+        && echo "COPIED SUCCESSFULLY"
+      sync
+      sudo umount "$PWD/__tmp_volume"
+      zip -9 --junk-paths $FILE.zip $FILE
+      rm $FILE
     done
     rm -rf __tmp_volume
 fi
