+++ test/JavaTests/testdata/simple/compress.sh
@@ -19,6 +19,7 @@ CREATE_SIMPLE_XAR=n
 CREATE_SIMPLE_UDF=n
 CREATE_SIMPLE_WIM=n
 CREATE_SIMPLE_FAT=n
+CREATE_SIMPLE_NTFS=n
 
 if test $CREATE_SIMPLE_ARJ = y -o $CREATE_ALL = y ;then
     rm arj/*.arj
@@ -213,26 +214,48 @@ if test $CREATE_SIMPLE_FAT = y -o $CREATE_ALL = y ;then
     do
       for i in 1:80 2:80 3:40 4:40 5:40
       do
-          #for j in *.dat
-          #do
-              FILE="fat/simple${i:0:1}.dat.$sectorsPerCluster.fat"
-              echo ""
-              echo "----------------> Creating $FILE"
-              echo ""
-              dd if=/dev/zero "of=$FILE" count=${i:2} bs=1024 \
-                && mkfs.vfat -s $sectorsPerCluster "$FILE" \
-                && sudo mount "$FILE" __tmp_volume \
-                && sudo cp simple${i:0:1}.dat __tmp_volume/ \
-                && ls -l __tmp_volume \
-                && echo "COPIED SUCCESSFULLY"
-              sync
-              sudo umount "$PWD/__tmp_volume"
-              zip -9 --junk-paths $FILE.zip $FILE
-              rm $FILE
-              #cp $j /$IMAGEX_SUBST_DRIVE/
-              #cmd /c "imagex /capture $IMAGEX_SUBST_DRIVE: wim/$j.${i:0:1}.wim Simple_test__$j /compress ${i:2}"
-          #done
+        FILE="fat/simple${i:0:1}.dat.$sectorsPerCluster.fat"
+        echo ""
+        echo "----------------> Creating $FILE"
+        echo ""
+        dd if=/dev/zero "of=$FILE" count=${i:2} bs=1024 \
+          && mkfs.vfat -s $sectorsPerCluster "$FILE" \
+          && sudo mount "$FILE" __tmp_volume \
+          && sudo cp simple${i:0:1}.dat __tmp_volume/ \
+          && ls -l __tmp_volume \
+          && echo "COPIED SUCCESSFULLY"
+        sync
+        sudo umount "$PWD/__tmp_volume"
+        zip -9 --junk-paths $FILE.zip $FILE
+        rm $FILE
       done
+    done
+    rm -rf __tmp_volume
+fi
+
+if test $CREATE_SIMPLE_NTFS = y -o $CREATE_ALL = y ;then
+    mkdir -p ntfs
+    rm -rf __tmp_volume
+    mkdir -p __tmp_volume
+    rm -f ntfs/*.ntfs
+    rm -f ntfs/*.zip
+
+    for i in 1:1025 2:1025 3:1025 4:1025
+    do
+      FILE="ntfs/simple${i:0:1}.dat.1.ntfs"
+      echo ""
+      echo "----------------> Creating $FILE"
+      echo ""
+      dd if=/dev/zero "of=$FILE" count=${i:2} bs=1024 \
+        && mkfs.ntfs -F "$FILE" \
+        && sudo mount "$FILE" __tmp_volume \
+        && sudo cp simple${i:0:1}.dat __tmp_volume/ \
+        && ls -l __tmp_volume \
+        && echo "COPIED SUCCESSFULLY"
+      sync
+      sudo umount "$PWD/__tmp_volume"
+      zip -9 --junk-paths $FILE.zip $FILE
+      rm $FILE
     done
     rm -rf __tmp_volume
 fi
